delete lots of old timeseries / latest / combined data code, likely breaking lots of tests. still trying to debug change in new_cases.
