+++ libs/datasets/data_source.py
@@ -1,14 +1,10 @@
-from itertools import chain
 from typing import Type, Optional
 
 import pandas as pd
 
-from covidactnow.datapublic.common_fields import CommonFields
-
 from libs.datasets.timeseries import MultiRegionDataset
 from libs.datasets.timeseries import TimeseriesDataset
 from libs.datasets.latest_values_dataset import LatestValuesDataset
-from libs.datasets.dataset_utils import AggregationLevel
 from functools import lru_cache
 
 
@@ -40,18 +36,6 @@ class DataSource(object):
         self.data = data
         self.provenance = provenance
 
-    @property
-    def state_data(self) -> pd.DataFrame:
-        """Returns a new BedsDataset containing only state data."""
-        is_state = self.data[CommonFields.AGGREGATE_LEVEL] == AggregationLevel.STATE.value
-        return self.data[is_state]
-
-    @property
-    def county_data(self) -> pd.DataFrame:
-        """Returns a new BedsDataset containing only county data."""
-        is_county = self.data[CommonFields.AGGREGATE_LEVEL] == AggregationLevel.COUNTY.value
-        return self.data[is_county]
-
     @classmethod
     def local(cls) -> "DataSource":
         """Builds data from local covid-public-data github repo.
@@ -59,16 +43,6 @@ class DataSource(object):
         Returns: Instantiated class with data loaded.
         """
         raise NotImplementedError("Subclass must implement")
-
-    @lru_cache(None)
-    def beds(self) -> LatestValuesDataset:
-        """Builds generic beds dataset"""
-        return self.latest_values()
-
-    @lru_cache(None)
-    def population(self) -> LatestValuesDataset:
-        """Builds generic beds dataset"""
-        return self.latest_values()
 
     @lru_cache(None)
     def multi_region_dataset(self) -> MultiRegionDataset:
@@ -88,28 +62,6 @@ class DataSource(object):
 
         raise ValueError("Unexpected index fields")
 
-    @lru_cache(None)
-    def timeseries(self) -> TimeseriesDataset:
-        """Build TimeseriesDataset from this data source."""
-        if set(self.INDEX_FIELD_MAP.keys()) != set(TimeseriesDataset.INDEX_FIELDS):
-            raise ValueError("Index fields must match")
-
-        return TimeseriesDataset.from_source(
-            self, fill_missing_state=self.FILL_MISSING_STATE_LEVEL_DATA
-        )
-
-    @lru_cache(None)
-    def latest_values(self) -> LatestValuesDataset:
-        if set(self.INDEX_FIELD_MAP.keys()) == set(TimeseriesDataset.INDEX_FIELDS):
-            return self.timeseries().latest_values_object()
-
-        if set(self.INDEX_FIELD_MAP.keys()) != set(LatestValuesDataset.INDEX_FIELDS):
-            raise ValueError("Index fields must match")
-
-        return LatestValuesDataset.from_source(
-            self, fill_missing_state=self.FILL_MISSING_STATE_LEVEL_DATA
-        )
-
     @classmethod
     def _rename_to_common_fields(cls: Type["DataSource"], df: pd.DataFrame) -> pd.DataFrame:
         """Returns a copy of the DataFrame with only common columns in the class field maps."""
@@ -117,10 +69,3 @@ class DataSource(object):
         to_common_fields = {value: key for key, value in all_fields_map.items()}
         final_columns = to_common_fields.values()
         return df.rename(columns=to_common_fields)[final_columns]
-
-    @classmethod
-    def _drop_unlisted_fields(cls: Type["DataSource"], df: pd.DataFrame) -> pd.DataFrame:
-        """Returns a copy of the DataFrame with columns not in the class field maps dropped."""
-        # Use pd.unique to preserve order, unlike making a set.
-        all_keys = pd.unique(list(chain(cls.INDEX_FIELD_MAP.keys(), cls.COMMON_FIELD_MAP.keys())))
-        return df[all_keys]
